refactor(TableBoar): to improve code structure and readability
